Reformat slack message + rename folder :)
